Refectoring PasswordFinder.py to be more modular, accept additional parsing use cases
